Add files via upload
So, problem 1 (ie the most obvious) - I can't get the footer to the bottom of the page. It's the right width, colour etc. I've seen this problem online and tried creating a container to separate the header and footer with no luck. Hoping there is an easy fix to this???

problem 2) The buttons are right at the top and whenever I add margins or padding it affects the dropdown menu by making the buttons way too big and then they don't hover correctly etc. 

problem 3) images are missing for the logo (just Car Rental text, figured it's easy to put it as an image with # so at least the positioning is fine and then we just need to upload a file), the sign in and the flag. At the moment I don't want to mess things up more so if someone knows how to do this then please help/take over.